refactor: split massive handlers.rs into domain modules and use Axum sub-routers
