+++ templates/add_fuel_entry.html
@@ -26,7 +26,7 @@
                    name="q"
                    placeholder="Type to search stations..."
                    autocomplete="off"
-                   hx-get="/htmx/stations/search"
+                   hx-get="/stations/htmx/search"
                    hx-trigger="keyup changed delay:200ms, focus"
                    hx-target="#station_results"
                    hx-indicator="#station-spinner"
