+++ templates/edit_fuel_entry.html
@@ -22,7 +22,7 @@
                    name="q"
                    placeholder="Type to search stations..."
                    autocomplete="off"
-                   hx-get="/htmx/stations/search"
+                   hx-get="/stations/htmx/search"
                    hx-trigger="keyup changed delay:200ms, focus"
                    hx-target="#station_results"
                    hx-indicator="#station-spinner"
@@ -62,7 +62,8 @@
             <button type="button"
                     class="btn btn-outline btn-danger"
                     style="margin-left: auto;"
-                    hx-delete="/htmx/entries/{{ entry.id }}"
+                    hx-delete="/fuel-entries/htmx/{{ entry.id }}"
+
                     hx-confirm="Are you sure you want to delete this fuel entry?"
                     hx-target="body"
                     hx-swap="none"
