+++ templates/fuel_entries.html
@@ -122,7 +122,16 @@
                         <td>€{{ "{:.2}"|format(entry.cost) }}</td>
                         <td>€{{ "{:.3}"|format(entry.cost / entry.litres) }}</td>
                         <td style="text-align: right;">
-                            <a href="/fuel-entries/{{ entry.id }}/edit" class="btn btn-small btn-outline">Edit</a>
+                            <div style="display: flex; gap: 0.25rem; justify-content: flex-end;">
+                                <a href="/fuel-entries/{{ entry.id }}/edit" class="btn btn-small btn-outline">Edit</a>
+                                <button class="btn btn-small btn-outline btn-danger"
+                                        hx-delete="/fuel-entries/htmx/{{ entry.id }}"
+                                        hx-target="#entry-{{ entry.id }}"
+                                        hx-swap="outerHTML"
+                                        hx-confirm="Are you sure you want to delete this entry?">
+                                    Del
+                                </button>
+                            </div>
                         </td>
                     </tr>
                     {% endfor %}
