+++ templates/vehicles.html
@@ -9,7 +9,7 @@
 </div>
 
 <div class="card">
-    <div id="vehicle-list" hx-get="/htmx/vehicles" hx-trigger="load">
+    <div id="vehicle-list" hx-get="/vehicles/htmx/list" hx-trigger="load">
         <p style="color: var(--text-muted);">Loading vehicles...</p>
     </div>
 </div>
